# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_nate_go-boilerplate_pkg_response.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
  internal_http_handler.AuthResponse:
    properties:
      token:
        $ref: '#/definitions/internal_http_handler.TokenResponse'
      user:
        $ref: '#/definitions/internal_http_handler.UserResponse'
    type: object
...
  internal_http_handler.LoginRequest:
    properties:
      identifier:
        example: user@example.com
        type: string
      password:
        example: securepassword123
        type: string
    required:
    - identifier
    - password
    type: object
...
  internal_http_handler.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  internal_http_handler.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword123
        minLength: 8
        type: string
      phone:
        example: "+1234567890"
        type: string
    required:
    - password
    type: object
  internal_http_handler.TokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsImlhdCI6MTUxNjIzOTAyMn0...
        type: string
      expires_in:
        example: 3600
        type: integer
      refresh_token:
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwidG9rZW5fdHlwZSI6InJlZnJlc2giLCJpYXQiOjE1MTYyMzkwMjJ9...
        type: string
    type: object
...
  internal_http_handler.UserResponse:
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone:
        example: "+1234567890"
        type: string
      role:
        example: user
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/nate/go-boilerplate
  description: |-
    A production-ready Go REST API boilerplate with authentication, notes, and user management.
    Features include JWT authentication, rate limiting, soft deletes, and Redis-based token storage.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go Boilerplate API
  version: "1.0"
paths:
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email/phone and password. Returns user info
        and authentication tokens.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/internal_http_handler.AuthResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      summary: Login to user account
      tags:
      - auth
  /api/v1/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the provided refresh token, logging out the user from
        the current session.
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Logged out successfully
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      summary: Logout from current session
      tags:
      - auth
  /api/v1/logout-all:
    post:
      consumes:
      - application/json
      description: Invalidate all refresh tokens for the authenticated user, logging
        out from all devices.
      produces:
      - application/json
      responses:
        "204":
          description: Logged out from all sessions successfully
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout from all sessions
      tags:
      - auth
...
  /api/v1/refresh:
    post:
      consumes:
      - application/json
      description: Get new access and refresh tokens using a valid refresh token.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens refreshed successfully
          schema:
            $ref: '#/definitions/internal_http_handler.TokenResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email or phone. Returns user info
        and authentication tokens.
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/internal_http_handler.AuthResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "409":
          description: Email or phone already registered
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"